plotCSLayer.m: VtxData_ci.mat path fixes: use global repo path, remove %
Previously plotCSLayer was set for calling it within the directory in
which it is stored. There is now a global variable REPODIR in the function
that enables saving VtxData_ci.mat even when calling the function
from anywhere, including outside of the repository.

NOTE: You must define REPODIR to be the full path to the CSGrid
repository within your calling scripts. REPODIR must be declared global
and "addpath(genpath(REPODIR));" must be applied.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>