Changes to example script compare_hp_to_classic.m to add features
This script allows easy comparison between outputs of GCHP runs and
GEOS-Chem classic runs. You can turn off either classic or GCHP in order
to compare results between multiple runs of a single model.

The requirement for comparison is that all outputs are contained in
separate run directories within a single testruns directory, and that
all run directories use a common prefix with unique identifier suffixes.

The key feature of this script is that it easily loads all output data
for selected species into data structures that are easy to use from
the matlab command line, or saved to a .mat file and loaded elsewhere
for data analysis.

Currently the metric printed out at the end for comparison is mean.
min and max are stored in the structure, and other metrics can easily
be added to print out.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>