Bug fix - wrong vertex data filename specification
The vertex data files used by plotCSLayer.m were being named incorrectly.
If a vertex data file is not found, it is generated and saved as
"VtxData_C#.mat" where # is the number of grid cells per cube
side. However, the code was missing a % sign, such that the first time
a grid was not found, the file "VtxData_Ci.mat" was generated. Subsequent
searches would always find the same file, resulting in the incorrect grid
data being read in. This bug has now been fixed.

Signed-off-by: Sebastian D. Eastham <seastham@fas.harvard.edu>