tempest_ewl branch: Add tempest xData-generation utility functions
Functions include:
	(1) flipCSFace
	(2) shiftLongitudes
	(3) swapCSFaces
	(4) transposeCSFace
	(5) readTempest

None of these have been tested, and some are blank except for header. I am
committing and pushing simply to have access remotely. The goal of these
functions to refactor working code currently in a single script.

A temp folder is also committed, containing the working script that
transforms tempest data into an xData struct array that can replicate a
GMAO tilefile for lat/lon -> cubed sphere regridding. The replication
is not exact, as there seems to be stretching in the tempest projection.
However, it is close enough to generate tilefiles of the needed resolutions
for general code testing (in this case, the motivation is to test my
drydep updates). The temp file also contains a tempest output file
(4x5 -> c48), and a gmao tilefile for testing.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>