Efficiency improvement in flipCSFaces
The routine flipCSFaces used a string comparison in an inner loop
to determine which flip to perform. This was inefficient due to
the fact that the string does not change. The string comparison
has now been moved outside and is used to determine two logicals,
which are used instead of the string comparison. Runtime when
processing a large cubed-sphere grid has been cut from > 30 s
to < 1 s.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>