Reduced I/O required when reading Tempest NetCDF files
Determining the properties of a rectangular grid in a Tempest
NetCDF file was dominating overall read time, due to the large amount
of data involved in the grid descriptors. This read time has been
reduced dramatically by taking advantage of some known properties
of the grid, such as assuming constant spacing in longitude. To
allow for the reduced rigor of these tests, the tolerance on the
test for date-centered/date-edged grids has been increased, but
the routines should remain accurate for grid spacings coarser than
1 meter.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>