Workaround for plotCSLayer not showing titles
Previously, the axes in plotCSLayer had their "visible" property
set to "off". This allowed the map axes to show up without a
large ugly rectangle around them. However, this also prevented
other axes objects, such as titles, from showing up. To get around
this, visible is now left "on" but the background box is disabled
piecemeal. The visible effect is the same but now features such
as the axes title can be used.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>