Add fix in readTempest.m to catch if lat/lon position improperly read
Determining PC or PE and DC or DE from tempest output data is not robust.
As a stopgap solution the determined positioning is compared with the
position strings in the input tempest netCDF filename. If there is a
discrepancy then the file string prevails and a warning message is printed.
If the position strings are not in the tempest netCDF filename then no
correction will take place and a warning is printed.

Also in this commit: update example script for creating a binary
tilefile from a tempest netCDF file so that it is compatible with
past updates to functions used within it.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>